requirements.txt
asgiref==3.8.1
blinker==1.8.2
click==8.1.7
Django==5.0.7
django-environ==0.11.2
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
psycopg2-binary==2.9.9
python-dotenv==1.0.1
sqlparse==0.5.1
typing_extensions==4.12.2
Werkzeug==3.0.3
